Mark APEXes that changed version during this boot
Some parts of the platform (e.g. PackageManager or ART) are interested
in knowing whether an APEX was updated during this boot. In case of
PackageManager this information is required to drop the cache of the
apks inside the updated APEX.

This change adds activeApexChanged field to ApexInfo.aidl. For a
given APEX in the given boot sequence this field will be true in one of
the following scenarios:

1. This APEX was updated during this boot.
2. This APEX was decompressed during this boot.
3. This APEX failed to be activated, and a pre-installed counterpart of
   it needed to be activated instead.

Bug: 225435110
Test: atest ApexTestCases
Change-Id: I5f2b7985f22abd07d41ce57f1a8fa3e38a2bd28d