Fix GetFactoryPackages() to use ApexFileRepository
/vendor partition can have multiple apexes installed for the same
package. ApexFileRepository scans the partition and choose only one of
them according to the system property.

GetFactoryFackages() scanning the directory will return the wrong list
of ApexFiles when there's such multi-installed apexes.

Now, GetFactoryFackages() uses the list of pre-installed apexes from
ApexFileRepository and it will skip APEXes which compose the
multi-install apexes but not chosen.

Bug: 230105395
Test: ApexTestCases
Change-Id: I54d415b67da4604ecad8d5c170c16f18f6b745ca