Increase the threshold for updatable-media's size
The new size is 780_824 (using 800_000 for the threshold to
account for small variations that could cause flakiness in
the test).

Bug: 213598310
Ignore-AOSP-First: The tested changes happen in internal. If we make these changes in AOSP, we would break presubmit in AOSP, which would allow a much larger updatable-media jar than intended (and the issue would be detected once automerger makes the change in internal master). So each branch should probably have its own threshold for presubmit to work.
Test: atest media_e2e_tests
Change-Id: I66117fb915e8c534aa866429292aa6fd7de9198a