Updating default version in tests
Instead of removing the version completely from the manifest we'll rely
on setting it to 0 as a way of marking it as default. This is done
because the release tooling is looking for a version to code to replace
during the build.

Test: presubmit
Bug: 231691643
Change-Id: I3cc32771b6ee45ced9f1e9e0b43b83fe3f8df39a
(cherry picked from commit 0a72792b23a4b35d0d8eff01ce64f8930208c127)
Merged-In: I3cc32771b6ee45ced9f1e9e0b43b83fe3f8df39a