Re-enable apex compression-related tests
Asserting on the equality of the decompressed APEX and the original
APEX, since they were built as 2 separate APEXes and hence have
different notices.

Instead in apex_file_test we assert on some properties of the
decompressed APEX that we can about, and for apexd_test we just don't
care about the exact equality, since the ValidateDecompressedApex will
already do the verification.

Test: atest ApexTestCases
Bug: 230030476
Change-Id: Ie66f270de39081da4520d55e969cddc5a63e279f