Configure tag depending on how apexd was launched
There are multiple different ways an apexd binary can be started
(bootstrap, snapshotde, vm, otachroot, unmount-all, etc.), having a
single log tag makes it a tad harder to figure out what use case the
apexd was started with.

This change configures the default tag to be different based on the use
case, which should make debugging a little bit easier.

Bug: 245284604
Test: boot
Test: check logs
Change-Id: I57319d5c877d30f5334e0987463a6a8a854732d0