Add dist entry for v2_unsigned_apk_container
com.android.apex.cts.shim.v2_unsigned_apk_container.apex is getting
copied to the dist directory as the module name without the apex
extension  and thus not getting updated with any of the recent prebuilt
drops.  That also causes it to be missing when attempting to add
prebuilts for riscv64.  Add a dist entry that sets the output name with
the apex extension.

Test: TARGET_BUILD_APPS=com.android.apex.cts.shim.v2_unsigned_apk_container m apps_only dist
Change-Id: I2cb25b821bb4070c3bd7bccd5b01c5921c301b82