Make sure to pre-populate uevent for APEXes
apexd has created empty dm devices early in the boot process in order to
minimize the time to wait for the corresponding /dev nodes when the
APEXes have to be mounted.

However, since kernel >= 5.15, the behavior has changed; the uevent is
not generated unless table is loaded.

So, this change switches to the new function `CreatePlaceholderDevice`
in libdm which ensures that uevent is created by loading an empty table
as well.

Bug: 286176029
Test: add some logs to the dm-ioctl.c in the kernel. Build CF and
observe the boot logs. It shows something like the below right after the
bootstrap apex.

06-22 01:42:23.399     0     0 D kobject : 'dm-17' (00000000022cead8): kobject_uevent_env
06-22 01:42:23.399     0     0 D kobject : 'dm-17' (00000000022cead8): kobject_uevent_env: uevent_suppress caused the event to drop!
06-22 01:42:23.399     0     0 D kobject : 'queue' (00000000c7b5b48a): kobject_uevent_env
06-22 01:42:23.399     0     0 D kobject : 'queue' (00000000c7b5b48a): kobject_uevent_env: filter function caused the event to drop!
06-22 01:42:23.399     0     0 D kobject : '254:17' (00000000750ea220): kobject_uevent_env
06-22 01:42:23.399     0     0 D kobject : 'dm-17' (00000000022cead8): kobject_uevent_env

Change-Id: I96338c46524470aaedc823e2d8c7c9e061572a6f