Fix flake of ApexTestCases
setns() may not work when a process is multithreaded. This causes the
test flakiness because MountNamespaceRestorer fails to restore the mount
namespace and other tests start in a wrong mount namespace.

With local tests, atest ApexTestCases --iteration 100
--test-filter=ApexdMountTest.OnStartInVmModeFailsWithWrongRootDigest:ApexdTestUtilsTest.SetUpApexTestEnvironment
shows 7% failure rate.

This could be fixed by making apexd back to single threaded. But the
current multithreaded APEX activation will be slowed then.

Instead, in this change, setns() is called in a loop. In general, this
is a horrible solution, but it's acceptable because in most cases,
setns() just succeeds and make tests more reliable. In worst case, tests
will fail with obvious error messages with timeout anyway.

With the fix, tests pass 100%.

Bug: 286947345
Test: atest ApexTestCases --rerun-until-failure 100
Change-Id: Ie382123cc9fd95d55770f2c66bac87a6c046fa84